+++ core/parser/src/main/java/com/blazebit/persistence/parser/SimpleQueryGenerator.java
@@ -548,6 +548,10 @@ public class SimpleQueryGenerator implements Expression.Visitor {
 
     @Override
     public void visit(FunctionExpression expression) {
+        if (expression.getRealArgument() != -1) {
+            expression.getExpressions().get(expression.getRealArgument()).accept(this);
+            return;
+        }
         BooleanLiteralRenderingContext oldBooleanLiteralRenderingContext = setBooleanLiteralRenderingContext(BooleanLiteralRenderingContext.PLAIN);
         ParameterRenderingMode oldParameterRenderingMode = setParameterRenderingMode(ParameterRenderingMode.PLACEHOLDER);
         List<Expression> expressions = expression.getExpressions();