+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Not yet released
 * Allow to control embedding of id query into pagination object query and enable by default if DBMS allows it
 * Allow to control embedding of count query into pagination object/id query and enable by default if possible
 * Implement option for inlining CTEs/Subquery in from clause and add lateral join support
+* Use `JpqlFunction` for `NULL` literal expressions
 
 ### Bug fixes